feat(ui-keylight): makes light update optimistic
UI is directly updated after the light values have been changed in
the ui. This solves the bug that the user had to wait for on the result
of the backend to set a new value.

In case the update is not successfull, it will be overwritten by the next
updated from the backend.