feat(go-KeylightCharm-GroupScenes): fixes filtering bug
Fixes but that filtering and applying scenes did not work properly.
The keystrokes were not passed properly to the table when in filter mode due
to keybinding conflicts between the main group details and scenes page.

In addition enabling a scene did not work properly when in filtered mode due to an
index missmatch.
The matching is now done via the name and not the index. Id is not possible as the
list does not support generics and only returns the name (FilterValue).